+++ src/views/templates/edit.php
@@ -215,7 +215,7 @@
 		<div class="custom_buttons">
 			<!-- ko foreach: actions -->
 				<!-- ko if: hasPermission && $root.actionPermissions[name] !== false -->
-					<input type="button" data-bind="click: function(){$root.customAction(name, messages)}, value: title,
+					<input type="button" data-bind="click: function(){$root.customAction(name, messages, confirmation)}, value: title,
 																	attr: {disabled: $root.freezeForm() || $root.freezeActions()}" />
 				<!-- /ko -->
 			<!-- /ko -->