+++ changelog.md
@@ -1,5 +1,6 @@
 ## Changelog
 
+- You can now supply a 'confirmation' string option to your custom actions which will require a confirmation from the admin user before the action can go through.
 - You can now specify an options_sort_field and an options_sort_direction for relationship fields that use accessors as name fields are require ordering on something other than the name_field
 - Bugfix: The model results no longer require an ajax load on pageload
 - Bugfix: Table prefixes are now taken into consideration